fix: Resolve all remaining compilation errors
Fixed 4 critical issues:

1. Removed non-existent Volo.Payment.Stripe package
   - Package doesn't exist in ABP 9.3.6
   - Removed from HttpApi.Host.csproj
   - Removed PaymentStripeModule dependency
   - Payment module works without gateway-specific packages

2. Fixed type conversion in OrderPaymentService
   - PaymentRequestProductCreateDto expects float, not decimal
   - Cast order.TotalAmount to float for UnitPrice and TotalPrice

3. Fixed CurrentTenant in test file
   - Added ICurrentTenant injection
   - Replaced all CurrentTenant.Id with _currentTenant.Id

4. Fixed QuestPDF Text() method chaining in InvoiceService
   - Text(lambda) doesn't return chainable object
   - Use DefaultTextStyle() inside lambda instead

All errors should now be resolved!

+++ src/Masroof.Ecommerce.Application/Payments/OrderPaymentService.cs
@@ -56,9 +56,9 @@ public class OrderPaymentService : ApplicationService
                 {
                     Code = order.OrderNumber,
                     Name = $"Order {order.OrderNumber}",
-                    UnitPrice = order.TotalAmount,
+                    UnitPrice = (float)order.TotalAmount,
                     Count = 1,
-                    TotalPrice = order.TotalAmount
+                    TotalPrice = (float)order.TotalAmount
                 }
             }
         };
